Renames Node to Path, create Djikstra namespace
I makes more sense to think of each step along a chain of visited links
as something more of a path: a linked-list-esque object that's only
got knowledge of who came before it.

This commit also adds add'l functionality to the now-christened Path
object, adds extra tests, and removes the Faker gem (which is no longer
being used in this test). May decide to use it again later, may not.