Create Djikstra Runner
This is kind of a god class, and for that reason, I kind of don't like it.
This object is in the Djikstra Namespace, but it's more of a Djikstra-Wiki
hybrid. It needs implmentation detail knowledge from both worlds to do its
job. I tried to extract all of the wiki knowledge and put it elsewhere, but
at a certain point I hit a wall.

The runner needs to know about pages to use the gateway. Simple as that.

In later commits, I may try to remove more wiki logic from this class. There's
a chance that I won't though: there's no such thing as a perfect project.